requirements.txt
crew
crewai-tools
python-dotenv
requests
sqlite3
datetime
